+++ include/tachimawari/control/sdk/module/dynamixel_sdk.hpp
@@ -51,16 +51,16 @@ public:
 
   bool ping(uint8_t id) override;
 
-  bool write_packet(uint8_t id, uint8_t address, int value, int data_length = 1) override;
+  bool write_packet(uint8_t id, uint16_t address, int value, int data_length = 1) override;
 
-  int read_packet(uint8_t id, uint8_t address, int data_length = 1) override;
+  int read_packet(uint8_t id, uint16_t address, int data_length = 1) override;
 
   bool sync_write_packet(const std::vector<joint::Joint> & joints, bool with_pid = false) override;
 
   bool send_bulk_read_packet() override;
   bool add_default_bulk_read_packet() override;
   int get_data(uint8_t id, uint16_t address, int data_lenghth = 1) override;
-  int get_bulk_data(uint8_t id, uint8_t address, int data_length = 1) override;
+  int get_bulk_data(uint8_t id, uint16_t address, int data_length = 1) override;
 
 private:
   dynamixel::PortHandler * port_handler;
